FileHelper removed destinationOptions

+++ src/helpers/FileHelper.php
@@ -17,7 +17,6 @@ class FileHelper extends \yii\helpers\FileHelper
      * @param string $source
      * @param string $destination
      * @param array $options настройки поиска исходной папки  [[FileHelper::findFiles]].
-     * атрибут destinationOptions настройки папки назначения
      * может быть указан атрибут delete, если false  запрещает  удалять файлы
      * @param callable $onProcess
      * @throws \Exception
@@ -29,8 +28,7 @@ class FileHelper extends \yii\helpers\FileHelper
 
         $deleteFiles = ArrayHelper::remove($options, 'delete', true);
 
-        list($new, $updated, $deleted) = self::compareDirectories($source, $destination, $options,
-            ArrayHelper::remove($options, 'destinationOptions', []));
+        list($new, $updated, $deleted) = self::compareDirectories($source, $destination, $options);
 
 
         foreach (array_merge($new, $updated) as $file) {
@@ -45,7 +43,7 @@ class FileHelper extends \yii\helpers\FileHelper
             }
         }
 
-        if ($deleteFiles && count($deleted) > 0) {
+        if ($deleteFiles && count($deleted) > 0 && is_dir($destination)) {
             foreach ($deleted as $file) {
                 self::unlink($destination . $file);
 
@@ -76,21 +74,20 @@ class FileHelper extends \yii\helpers\FileHelper
      * Сравнение файлов в каталогах
      * @param string $source
      * @param string $destination
-     * @param array $sourceOptions настройки поиска каталога источника [[FileHelper::findFiles]]
-     * @param array $destinationOptions
+     * @param array $options
      * @return array list [$new, $updated, $deleted] массива список новых, измененных, удаленных файлов
      * @throws \Exception
      */
-    public static function compareDirectories($source, $destination, $sourceOptions = [], $destinationOptions = [])
+    public static function compareDirectories($source, $destination, $options = [])
     {
         $source = \Yii::getAlias($source);
         $destination = \Yii::getAlias($destination);
 
-        $sourceList = self::findFilesRelative($source, $sourceOptions);
-        $destinationList = is_dir($destination) ? self::findFilesRelative($destination, $destinationOptions) : [];
+        $sourceList = self::findFilesRelative($source, $options);
+        $destinationList = is_dir($destination) ? self::findFilesRelative($destination, $options) : [];
 
         $updated = [];
-        $compareTime = ArrayHelper::getValue($sourceOptions, 'compareTime', true);
+        $compareTime = ArrayHelper::getValue($options, 'compareTime', true);
 
         foreach (array_intersect($sourceList, $destinationList) as $file) {
             if (filesize($source . $file) !== filesize($destination . $file)) {
